added deblur shader and adjusted gba fullscreen
genesis and neogeo could appear blurry as a side effect of the tvout shader to blend dithering and reduce rainbowing, with this deblur shader they look better without introducing severe artifacts.
the genesis shader preset was also adjusted to include 32x games that use a different emulator (picodrive instead of genesis plus gx).
it's also applied to pc and pc98 because pc games can become unreadable and this helps a little.
doom and quake are deblured too and also changed them to use the pc shader preset because they don't seem to need the gamma adjust and the black levels were being crushed.
on a last note, gba fullscreen was adjusted to look as big as the fullscreen option of the gameboy player.